uv-docker-example/pyproject.toml
[project]
name = "uv-docker-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "fastapi[standard]>=0.112.2",
    "flower>=2.0.1",
    "openai>=1.85.0",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.9.1",
    "pymysql>=1.1.1",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "ttkbootstrap>=1.13.10",
    "requests>=2.32.4",
     "PyQt5>=5.15.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
hello = "uv_docker_example:hello"

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.2",
    "fastapi-cli>=0.0.5",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pymupdf>=1.26.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
